+++ includes/trophies-magtricky.php
@@ -1,0 +1,23 @@
+<h2 class="subtitled">MagTricky</h2>
+<p class="subtitle">The magnetic trophy for the listeners.</p>
+
+<div class="image-grid img2">
+	<img src="images/about/magtricky-1.jpg" />
+	<img src="images/about/magtricky-2.jpg" />
+</div>
+
+
+<p>
+	For listeners at home there’s the <a href="https://www.magtricky.com" target="_blank">MagTricky</a>, the <a href="/billof/annual-2022#rule120">official
+	way</a> to keep up with who has won the Rickies. A 10cm (4”) flat magnet featuring the majestic Connected globe artwork
+	allows the passionistas to track who currently holds Chairman titles with small cylindrical magnets. Because magnets
+	are awesome.
+</p>
+<p>
+	Unlike the official Tricky, there are two smaller silver 5mm magnets for when the titles are dispersed, and one
+	larger golden 7mm magnet for the consolidated chairman, or “<a href="/billof/keynote-jun-2020#rule88">Mega Chairman</a>”.
+</p>
+<?= $interactive_magtricky ?>
+<p>
+	For those that did not get the chance to buy the MagTricky, the above is an interactive digital MagTricky. Just click on the cities and the magnets will update. On refresh the chairman magnets are reset to current states according to Rickies.co.
+</p>